Updates the README to to include changes in 1.0.3
## Core

In #16, it's mentioned that `onChange` as of `1.0.3` requires a 0-arity function and notifies when changes are being made. In order to access the sanitized data, `onData` has to be used which requires a 1-arity function. Furthermore, `holderId` was deprecated and `holder` has to be used.

The documentation didn't include the changes to `onChange`, `onData` and `holderId`. This PR changes the README example to include the changes of `onChange`,`onData` and `holderId`.